.github/skills/table-spec-extractor/scripts/requirements.txt
# Excel
openpyxl>=3.1

# PDF rendering (includes pdfium — no system poppler needed)
pypdfium2>=4.20

# PDF text extraction (optional, improves text quality)
pdfplumber>=0.11

# Table Transformer (DETR-based vision models)
torch>=2.1
torchvision>=0.16
transformers>=4.40

# Image processing
Pillow>=10.0

# Neo4j driver
neo4j>=5.10
